fixed the conflicts vxrcel caused
