+++ src/app/(Public-Routes)/layout.tsx
@@ -1,8 +1,10 @@
 import type { Metadata } from "next"
 import { Inter, Roboto_Condensed } from "next/font/google"
-import "./globals.scss"
-import Navbar from "@/components/navbar/Navbar"
+import "../globals.scss"
+// import Navbar from "@/components/navbar/Navbar"
 import Footer from "@/components/footer/Footer"
+import Navbar from "@/components/navbar/Navbar"
+import { NavContextProvider } from "@/context/NavContext"
 
 export default function RootLayout({
   children,
@@ -10,10 +12,10 @@ export default function RootLayout({
   children: React.ReactNode
 }) {
   return (
-    <>
+    <NavContextProvider>
       <Navbar />
       {children}
       <Footer />
-    </>
+    </NavContextProvider>
   )
 }
