+++ src/app/layout.tsx
@@ -2,7 +2,7 @@ import type { Metadata } from "next"
 import { Inter, Roboto_Condensed } from "next/font/google"
 import "./globals.scss"
 // import Navbar from "@/components/navbar/Navbar"
-import Navbar from "@/components/Navbar copy/Navbar"
+import Navbar from "@/components/navbar/Navbar"
 import Footer from "@/components/footer/Footer"
 import GotoTop from "@/components/GotoTop"
 import { NavContextProvider } from "@/context/NavContext"
